Adding tests for constructor annotations
This tests @ConserviativeFlow on constructors for parameters to and return types of constructors.

This also tests explicit return type annotations of constructors.

The tests have currently been made to pass to not break builds. There are notes on how the tests should fail. After review the tests will be enabled to test fixes / find regressions.