Flow checker outputs missingAPI.astub.
This file contains a list of methods that need to be add to the
stub files. The file also contains a comment above the method with
the number of times the methods is used in the source code.

FlowChecker also prints to System.err how many methods need to be annotated.

TODO: Add this to the manual.